+++ IsekaiMod/Content/Features/IsekaiProtagonist/OverpoweredAbility/OverpoweredAbilitySelection.cs
@@ -45,10 +45,6 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.OverpoweredAbility
                 bp.SetName("Mythic Overpowered Ability");
                 bp.SetDescription("You use your mythic powers to gain an additional Overpowered Ability.");
                 bp.m_Icon = Icon_TrickFate;
-                bp.AddComponent<PrerequisiteFeature>(c => {
-                    c.Group = Prerequisite.GroupType.All;
-                    c.m_Feature = OverpoweredAbilitySelection.ToReference<BlueprintFeatureReference>();
-                });
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
                 bp.m_AllFeatures = new BlueprintFeatureReference[0];